Moved Post Component to its own folder
